Modularize, Rename to settings.yaml
