+++ ad-ip-acg/pyproject.toml
@@ -7,7 +7,7 @@ readme = "README.md"
 
 [tool.poetry.dependencies]
 python = "^3.11"
-boto3 = "^1.35.99"
+boto3 = "^1.36.0"
 
 
 [tool.poetry.group.dev.dependencies]
